Add About loader and extract hero actions
Introduce a dedicated about loader (app/routes/about/loader.ts) that provides leaders (with buildLeadersWithBios fallback to staticLeaders) and update About route/components to use AboutLoaderData. Extract hero actions fetching into a new server module (app/routes/home/hero-actions.server.ts) and wire it into the Home loader (app/routes/home/loader.ts) so HomeLoaderData now includes actions. Update hero components to use useLoaderData<HomeLoaderData>() instead of the root route loader. Remove hero-actions logic and the actions field from the navbar/root loader (app/routes/navbar/loader.tsx) to decouple responsibilities and simplify the root loader's payload.
