+++ app/routes/home/components/hero/desktop-features.component.tsx
@@ -1,10 +1,10 @@
 import { Icon } from '~/primitives/icon/icon';
 import { cn } from '~/lib/utils';
-import { Link, useRouteLoaderData } from 'react-router-dom';
-import { HeroAction, RootLoaderData } from '~/routes/navbar/loader';
+import { Link, useLoaderData } from 'react-router-dom';
+import { HeroAction, HomeLoaderData } from '~/routes/home/loader';
 
 export const DesktopFeaturedItems = () => {
-  const { actions } = useRouteLoaderData('root') as RootLoaderData;
+  const { actions } = useLoaderData<HomeLoaderData>();
   if (!actions) {
     return null;
   }
